### 2021-02-05 #3013 @whedon
[PRE REVIEW]: ARDL: An R package for the analysis of level relationships
**Submitting author:** @Natsiopoulos (<a href="http://orcid.org/0000-0003-1180-2984">Kleanthis Natsiopoulos</a>)
**Repository:** <a href="https://github.com/Natsiopoulos/ARDL" target ="_blank">https://github.com/Natsiopoulos/ARDL</a>
**Version:** v0.1.1
