### 2021-02-05 #3013 @kyleniemeyer
Hello @Natsiopoulos, thanks for your interest in JOSS. We use this pre-review issue to identify an editor and reviewers, and also resolve any obvious issues.

Your repository is currently missing an open-source license file; I see in the DESCRIPTION file that you use the GPL-3 license, so please add this to the repo at the top level.

Also, your paper needs a Statement of Need section, which is a [required element](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of the paper.
