+++ JOSS-paper/paper.bib
@@ -10,6 +10,7 @@
     title = {{dLagM}: An {R} package for distributed lag models and {ARDL} bounds testing},
     author = {Haydar Demirhan},
     journal = {PLoS ONE},
+    doi = {10.1371/journal.pone.0228812},
     volume = {15},
     number = {2},
     pages = {e0228812},
@@ -39,4 +40,15 @@
   pages = {289--326},
   volume = {16},
   year = {2001}
+}
+
+@article{Qiu2021,
+title = {{TOURIST ARRIVAL FORECAST AMID COVID}-19: A perspective from the {Asia} and {Pacific} team},
+journal = {Annals of Tourism Research},
+pages = {103155},
+year = {2021},
+issn = {0160-7383},
+doi = {https://doi.org/10.1016/j.annals.2021.103155},
+url = {https://www.sciencedirect.com/science/article/pii/S0160738321000177},
+author = {Richard T.R. Qiu and Doris Chenguang Wu and Vincent Dropsy and Sylvain Petit and Stephen Pratt and Yasuo Ohe}
 }
