### 2021-02-06 #3013 @Natsiopoulos
Hello @kyleniemeyer, thank you for accepting our submission.

On behalf of my coauthor, I have made the following changes.
- Added the license file.
- Added the Statement of Need section. The relative information was already there but as part of the summary.
- Added a reference about a recent publication that uses the software.

Thank you
