### 2021-03-11 #3013 @kyleniemeyer
Hello @Natsiopoulos, thanks for your patience. Unfortunately, many in our editorial board are still at reduced capacity right now, but submissions have not really slowed.

I will be looking for an alternative editor, but it may take a bit longer still as we work through the backlog.
