### 2021-05-06 #3013 @sbenthall
@Natsiopoulos are you suggesting that @alexpghayes be asked to review the package for JOSS? In any case, of course you may take additional input from others and @alexpghayes is welcome to use the review issue to provide additional feedback. We will need to keep the official review process focused on the two formally chosen reviewers.
