### 2021-06-03 #3013 @ha0ye
> @ha0ye Thank you! We are a little backed up in the process because we need to find a second reviewer. But if you do your review soon, you can submit it once the second reviewer is found.
>
> Do you have the review checklist available?

Yep, I've reviewed for JOSS before, as well as used the rOpenSci checklist in a separate setting.
