### 2021-06-03 #3013 @Darthpathos
> @Darthpathos would you be able to review this package?

Hi @sbenthall - Unfortunately I don't anything about economics (I'm in healthcare).  Please let me know if you have any packages needing review for clinical trials, medical research, etc. :)
