### 2021-06-08 #3013 @feddelegrand7
Hello @sbenthall. Thank you for asking, unfortunately I don't have the "time series" background and  necessary skills to review the package.
