### 2021-07-09 #3013 @jacobsoj
Hi Sebastian,
Sorry for the delay. The email got filtered to a package that I don't often check. I can review this package if you still need me to.
Jill

Jill A. Jacobson, Ph.D. Office Phone  613-533-2847
Associate Professor, Social-Personality Program
Lab Phone 613-533-6000 x75417
Queen's University      FAX  613-533-2499
Department of Psychology        Email  ***@***.***
Kingston, ON  K7L 3N6
CANADA

http://www.queensu.ca/psychology/People/Faculty/Jill-Jacobson
