### 2021-07-15 #3496 @whedon
[REVIEW]: ARDL: An R package for the analysis of level relationships
**Submitting author:** @Natsiopoulos (<a href="http://orcid.org/0000-0003-1180-2984">Kleanthis Natsiopoulos</a>)
**Repository:** https://github.com/Natsiopoulos/ARDL
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
