### 2021-08-22 Natsiopoulos/ARDL#6 @ha0ye
[JOSS Review] Notes from @ha0ye
This includes all my notes for the unchecked boxes for my JOSS review, as well as additional suggestions for improvements.

(see comments on **statement of need** in "Software paper" below)

An example using the `denmark` included dataset is provided in the README. It is a bit long, but primarily shows the main functions of the package.

I don't know how useful the `Ease of use` section is at the bottom of the README. If it's an important draw for the package, then it should be higher up. Otherwise, it's not likely to be convincing after someone has already read through the example. You could consider expanding on both the example and the `Ease of use` section in a vignette. You could then provide advice in the README and perhaps `.onAttach()` for users to refer to the vignette for a detailed example.

Yes, in the "Breaking down the ARDL package" section of the software paper. This should be moved somewhere else, possibly a separate vignette describing the model objects of the package. I appreciate the authors documenting this, but this is specifically noted as not fitting the guidelines for the software paper.

The package is missing tests. Since there are only a few functions, a bit of dedicated effort to convert some of the examples into `testthat` tests would be helpful and not too much work, I think. See https://r-pkgs.org/tests.html for details.

No, but this is a relatively easy fix:
* add a `BugReports` field to DESCRIPTION, with link to https://github.com/Natsiopoulos/ARDL/issues
* add a CONTRIBUTING file to the github repo, rOpenSci has a template at https://github.com/ropensci/dotgithubfiles/blob/master/dotgithub/CONTRIBUTING.md

It is not clear with ARDL or ECM are, though it is implies that this package is implementing these approaches. Lines 13-17 of the Statement of Need could be effective at as a summary, if an additional sentence were added that lists the functionality of the package, along with references for these methods.

Currently, the statement of need repeats much the same information as the summary, and could be clearer about the problem being solved by the software and the audience. Here are suggestions on how to improve it:
* Explain *what* kind of flexibility is encapsulated in ARDL and ECM. (the summary mentions they are very flexible, which is fine and good for modeling approaches, but what kind of situation warrants this flexibility?)
* The software is designed to simplify the construction of these models (i.e. the user need only specify the order for the ARDL model). This should be emphasized as improving the accessibility of these modeling approaches for e.g. students and users who apply time series modeling, but are perhaps not experts in software, the mathematics of time series modeling, etc.
* The link to the manual and recent example do not really fit here. Instead, link to a vignette with examples, and move the Qiu et al. to the summary.

Mostly yes, but I think this would be improved by merging the content of the "Ease of use" section of the README here, as well as the "Conclusion" section of the software paper.

The language is acceptable, though some expansion of the modeling approaches for the time series novice would be appreciated.

Citations should be added that expand on the time series methods.

### Additional Notes
* I recommend adding `JOSS-paper` to `.Rbuildignore`,
* You have citation information in `inst/CITATION` already, so it'd be ideal to reduce duplication in the `.onAttach()` function. Something like `format(citation("ARDL"), bibtex = FALSE)` will generate much the same text.
* Consider adding a `CITATION.cff` file to the github root (and add this to `.Rbuildignore`, too) - https://citation-file-format.github.io/
* Consider creating a **pkgdown** documentation site. I ran it on my local clone of the repo without issue, and this will provide a nicer set of docs to link users to, e.g. from the README.
* I didn't poke deeply into the code, but `unlist(lapply(1:nrow(top_orders), FUN = function(z) sum(order2_back == top_orders[z,-ncol(top_orders)])))` is a bit awkward, and you might be able to get away with the built-in `matrix::colSums()` function or by using `apply()` over rows.
