### 2021-08-22 #3496 @ha0ye
My review notes are now in https://github.com/Natsiopoulos/ARDL/issues/6
