### 2021-10-04 #3496 @jacobsoj
My father died, so I have been out of the country. If you can wait another week(Oct. 11), I can finish this review. If not, then I withdraw. I apologize, but it has been a difficult time. It was under hospice care for 2,months.
Jill

Sent from my iPhone
