### 2021-12-15 #3496 @jacobsoj
Sorry, Sebastian, I kept hoping I could rally to get this done. I hope the other person can do it.

Could you withdraw my name/contact from future consideration for reviews? I use R a lot, but I rarely use GitHub and have to relearn it each time, which adds to the burden of doing these reviews.
Jill

Jill A. Jacobson, Ph.D. Office Phone  613-533-2847
Associate Professor, Social-Personality Program
Lab Phone 613-533-6000 x75417
Queen's University      FAX  613-533-2499
Department of Psychology        Email  ***@***.***
Kingston, ON  K7L 3N6
CANADA

http://www.queensu.ca/psychology/People/Faculty/Jill-Jacobson
