### 2022-01-20 #3496 @jessie-dotson
This package implements a model that I'm not familiar with, and am unlikely to use -- but it does look interesting and useful for other use cases.  If JOSS is ok with having a reviewer who is unlikely to use the package in the future then I'm happy to do a review.  (I should be able to complete in the next 2 weeks.)
