### 2022-01-27 #3496 @sbenthall
@jessie-dotson Yes, it's perfectly fine for reviewers to review packages that they are unlikely to use themselves.

Thank you for volunteering! I'll assign you as the second reviewer.
