### 2022-01-29 Natsiopoulos/ARDL#7 @jessie-dotson
Suggested changes to the JOSS article
A few typos and edits to consider for JOSS manuscript:

- line 19:  replace "amonth" with "among"
- line 41:  the "too" at the end of the sentence is unnecessary
- equations would be clearer if variables were defined
- lines 50 - 54:  Would be clearer if there were a clear explanation of what determined a "Case"
- equations 4-8:  ECT is not defined
- line 95:  include link to API documentation when discussing the API
