### 2022-01-29 #3496 @jessie-dotson
Suggested edits to JOSS paper in https://github.com/Natsiopoulos/ARDL/issues/7#issue-1117989915
