### 2022-02-10 Natsiopoulos/ARDL#8 @jessie-dotson
auto_ardl generates a warning when used with current version of tibble
When using the current version of tibble (3.1.6), the auto_ardl function produces a compatibility warning.  The warning below was generated by executing the example call in the usage documentation on github

`models <- auto_ardl(LRM ~ LRY + IBO + IDE, data=denmark, max_order=5)`

```
Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
Using compatibility `.name_repair`.
```
