### 2022-02-10 #3496 @jessie-dotson
Recommend fixing use of tibble to avoid compatibility warning.  Documented in https://github.com/Natsiopoulos/ARDL/issues/8#issue-1129251801
