### 2022-02-16 Natsiopoulos/ARDL#9 @jessie-dotson
JOSS review -- please add clear community guidelines
The JOSS review checklist includes the item below:

_Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_

Please add appropriate guidelines to the repository to meet JOSS guidelines.
