### 2022-02-16 #3496 @jessie-dotson
Community guidelines missing.  Please see https://github.com/Natsiopoulos/ARDL/issues/9
