### 2022-02-16 Natsiopoulos/ARDL#10 @jessie-dotson
JOSS Review -- tests to verify functionality needed
The JOSS review checklist includes the item below.

_Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?_

I didn't see any tests that would fulfill this criteria.  Please add either automated tests or outline manual steps to verify functionality.  (If the tests or process exists, please label them appropriately.)
