### 2022-02-16 #3496 @jessie-dotson
Tests to verify functionality missing.  Please see https://github.com/Natsiopoulos/ARDL/issues/10
