### 2022-02-16 #3496 @jessie-dotson
I've worked through the entire review checklist.  Outstanding items to address are documented in issues 7-10 opened in the packages repository and linked above.  After these are closed or addressed, please feel free to @ me so I'll know to come back and check off the changes.
