### 2022-06-28 79435de
Documentation update. Add DOI, references and citation for the publication in JAE
