+++ NEWS.md
@@ -1,3 +1,13 @@
+# ARDL 0.2.0
+
+### Documentation update
+
+* Add DOI and references of the publication where the package is used and 
+successfully replicates the results of the methodology it implements, in the 
+`DESCRIPTION` and `README.Rmd` files.
+
+* Properly changed the citation info in the `CITATION` and `ardl-package.R`files.
+
 # ARDL 0.1.1
 
 ### New features
