### 2022-08-05 #3496 @arfon
@Natsiopoulos – Arfon Smith, Editor in Chief of JOSS here. If we don't hear back from you in the next month (5th September 2022) we will assume this submission is abandoned and reject.
