### 2022-08-10 4cca518
Update of documentation and software paper for JOSS
