+++ JOSS-paper/paper.md
@@ -6,7 +6,7 @@ tags:
 - cointegration
 - econometrics
 - R
-date: "6 February 2021"
+date: "10 August 2022"
 output:
   word_document: default
   html_document:
@@ -28,11 +28,13 @@ affiliations:
 
 # Summary
 
-Autoregressive Distributed Lag (ARDL) and Error Correction Models (ECM) are widely used in various economic applications as they are very flexible. Also, these models are used in the context of cointegration analysis as a platform to test and analyze the levels (long-run) relationship between variables. One of the most popular such tests is the bounds test proposed by @Pesaran2001 which allows testing for cointegration while at the same time estimates the level relationship.
+Autoregressive Distributed Lag (ARDL) and Error Correction Models (ECM) [@Pesaran1999] are widely used in various economic applications as they are very flexible. Also, these models are used in the context of cointegration analysis as a platform to test and analyze the levels (long-run) relationship between variables. One of the most popular such tests is the bounds test proposed by @Pesaran2001 which allows testing for cointegration while at the same time estimates the level relationship. `ARDL` [@Natsiopoulos2022; @ardl-pkg] is an `R` package that aims to help users in the modeling process of ARDL and ECM and it also provides the tools towards the bounds test for cointegration.
 
 # Statement of need
 
-`ARDL` [@ardl-pkg] is an `R` package that aims to help users in the modeling process of ARDL and ECM and it also provides the tools towards the bounds test for cointegration. `ARDL` is implemented in such a way that researchers can use it as a full featured tool for this specific type of analysis and students of all levels can be aware of how each piece of code works through the analytical [manual](https://cran.r-project.org/web/packages/ARDL/ARDL.pdf) and the examples which cover every functionality of the package. A recent example is @Qiu2021 where the software was used to forecast tourist arrivals through an ARDL model, amonth other methods.
+`ARDL` is implemented in such a way that researchers can use it as a full featured tool for this specific type of analysis and students of all levels can be aware of how each piece of code works through the analytical [manual](https://cran.r-project.org/web/packages/ARDL/ARDL.pdf) and the examples which cover every functionality of the package. One of the greatest advantages of the package is that there is no need for the user to know the complex mathematical equations and write them by hand. Instead, the correct specification of the model is created automatically, just by providing the order of the ARDL. Some recent examples of publications that used the package are @baruah2022 and @Qiu2021.
+
+It is very important to note that the `ARDL` package is the first software which managed to fully replicate the original paper by @Pesaran2001, verifying the validity and accuracy of the package [@Natsiopoulos2022].
 
 # State of the field
 
@@ -42,11 +44,57 @@ On the other hand, `dLagM` and `dynamac` provide additional plotting functionali
 
[…]
