+++ JOSS-paper/paper.bib
@@ -6,6 +6,23 @@
     url = {https://CRAN.R-project.org/package=ARDL},
   }
 
+@article{baruah2022,
+  title={Investigating Commodity Price Relations across Wholesale Markets: The Case of Paddy in Chhattisgarh, India},
+  author={Baruah, Prerona},
+  journal={Indian Journal of Agricultural Economics},
+  volume={77},
+  number={1},
+  year={2022}
+}
+
+  @Manual{dynlm-pkg,
+    title = {{dynlm}: Dynamic Linear Regression},
+    author = {Achim Zeileis},
+    year = {2019},
+    note = {R package version 0.3-6},
+    url = {https://CRAN.R-project.org/package=dynlm},
+  }
+
 @Article{dlagm-pkg,
     title = {{dLagM}: An {R} package for distributed lag models and {ARDL} bounds testing},
     author = {Haydar Demirhan},
@@ -26,6 +43,46 @@
     url = {https://CRAN.R-project.org/package=dynamac},
 }
 
+@article{Johansen1990,
+  title = {{Maximum Likelihood Estimation and Inference on Cointegration – with Applications to the Demand for Money}},
+  author = {Johansen, S. and Juselius, K.},
+  journal = {Oxford Bulletin of Economics and Statistics},
+  doi = {https://doi.org/10.1111/j.1468-0084.1990.mp52002003.x},
+  number = {2},
+  pages = {169--210},
+  volume = {52},
+  year = {1990}
+}
+
+@article{Natsiopoulos2022,
+author = {Natsiopoulos, Kleanthis and Tzeremes, Nickolaos G.},
+title = {ARDL bounds test for cointegration: Replicating the Pesaran et al. (2001) results for the UK earnings equation using R},
+journal = {Journal of Applied Econometrics},
+volume = {n/a},
+number = {n/a},
+pages = {},
+keywords = {ARDL, bounds test, cointegration, earnings equation, R},
+doi = {https://doi.org/10.1002/jae.2919},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jae.2919},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jae.2919},
+abstract = {Summary This paper replicates the UK earnings equation using the autoregressive distributed lag (ARDL) modeling approach and the bounds test for cointegration by Pesaran et al. (Journal of Applied Econometrics, 2001, 16(3), 289–326). The findings from the narrow sense fully replicate the original results using the open-source language R and the ARDL package. In the wide sense replication, augmented data are employed, thus extending the end period from 1997:Q4 to 2019:Q4, using an alternative measure for union power. Adopting the new dataset, this study reinvestigates the UK earnings equation, thereby providing supporting evidence of a long-run relationship and reveals empirical findings about the long-run effects of productivity, unemployment, tax wedge, and union power on wages.}
+}
+
+@inbook{Pesaran1999, 
+place={Cambridge}, 
+series={Econometric Society Monographs}, 
+title={An Autoregressive Distributed-Lag Modelling Approach to Cointegration Analysis}, 
+DOI={10.1017/CCOL521633230.011},
+booktitle={Econometrics and Economic Theory in the 20th Century: The Ragnar Frisch Centennial Symposium},
+chapter = 11,
+publisher={Cambridge University Press},
+author={Pesaran, M. Hashem and Shin, Yongcheol},
+editor={Strøm, SteinarEditor}, 
+year={1999}, 
+pages={371–413},
+collection={Econometric Society Monographs}
+}
+
 @article{Pesaran2001,
[…]
