+++ R/ardl-package.R
@@ -13,14 +13,7 @@
 globalVariables(c("."))
 
 .onAttach <- function(libname, pkgname) {
-	# local_version = paste("R package version", utils::packageVersion('ARDL'))
-	# year <- sub("-.*", "", utils::packageDate('ARDL'))
-	# packageStartupMessage('To cite ARDL in publications use:
- #  \nKleanthis Natsiopoulos and Nickolaos Tzeremes (',
- #  year,
- #  '). ARDL: ARDL, ECM and Bounds-Test for Cointegration. ',
- #  local_version,
- #  '. https://CRAN.R-project.org/package=ARDL')
-    packageStartupMessage('To cite ARDL in publications use:
-        \nKleanthis Natsiopoulos and Nickolaos G. Tzeremes (2022). "ARDL bounds test for cointegration: Replicating the Pesaran et al. (2001) results for the UK earnings equation using R", Journal of Applied Econometrics, https://doi.org/10.1002/jae.2919')
+    packageStartupMessage(format(citation("ARDL"), bibtex = FALSE)[[1]],
+                      "\n\n",
+                      format(citation("ARDL"), bibtex = FALSE)[[2]])
 }
