+++ DESCRIPTION
@@ -9,7 +9,7 @@ Description: Creates complex autoregressive distributed lag (ARDL) models and
     equation. The validity and the accuracy of this package have been verified 
     by successfully replicating the results of Pesaran et al. (2001) in 
     Natsiopoulos and Tzeremes (2022) <doi:10.1002/jae.2919>.
-Version: 0.2.0
+Version: 0.2.0.9000
 Authors@R: c(
   person("Kleanthis", "Natsiopoulos", 
   email = "klnatsio@gmail.com", role = c("aut", "cre"), 
@@ -17,8 +17,10 @@ Authors@R: c(
   person("Nickolaos", "Tzeremes", 
   email = "bus9nt@econ.uth.gr", role = "aut",
   comment = c(ORCID = "0000-0002-6938-3404")))
+BugReports: https://github.com/Natsiopoulos/ARDL/issues
 License: GPL-3
 URL: https://github.com/Natsiopoulos/ARDL
+Video: https://www.youtube.com/watch?v=keDUwWBpHjU
 Encoding: UTF-8
 LazyData: true
 Depends: 

[…]
