### 2022-08-10 Natsiopoulos/ARDL#6 @Natsiopoulos
Hi @ha0ye
Thank you for the thorough review.
Here are our replies about how we treated each of your suggestions.

---

**Functionality documentation:**

**Ha0ye:** _"Yes, in the "Breaking down the ARDL package" section of the software paper."_
_"I appreciate the authors documenting this, but this is specifically noted as not fitting the guidelines for the software paper."_

**Answer:** We removed the detailed descriptions of the equations to align with the guidelines for the software paper. We kept only the equations for ARDL and UECM as these are the building blocks of the package and along with the bounds tests they from the core functionality of the software.

**Ha0ye:** _"This should be moved somewhere else, possibly a separate vignette describing the model objects of the package."_

**Answer:** Thank you for the suggestion. This information is already included in the manual.

---

**Automated tests:**

**Answer:** Thank you for suggesting using `testthat`, it is indeed something that we plan to do.
However, there are already excessive tests for every function, (most of them included in the internal function `parsers` as this is the workhorse behind every other function). These are not in an automated format though, but rather manual tests.

As manual tests are permitted, can this be considered as done?

---

**Community guidelines:**

**Answer:** Thank you for the suggestions and the useful links. `BugReports` and CONTRIBUTING have been added.

---

**Summary:**

**Ha0ye:** _"It is not clear with ARDL or ECM are, though it is implies that this package is implementing these approaches."_

**Answer:**
As these are widely known models in undergraduate level of various fields of studies (e.g. economics, statistics, environmental studies etc), we don't think that this should be overspecified. Also, ARDL is self defined by its name and we believe that a technical definition does not fit the purpose of this summary, as this is supposed to describe the "high-level functionality and purpose of the software for a diverse, non-specialist audience".
Although, we added a reference (Pesaran and Shin, 1999) which is the paper that introduces the ARDL and ECM as platforms for the cointegration test in the paper of Pesaran et al, (2001).
Please let us know if you think otherwise.

**Ha0ye:** _"...if an additional sentence were added that lists the functionality of the package, along with references for these methods"_

**Answer:** Following your suggestion, we added the first sentence of the section "Statement of need" in the summary. This describes the functionality of the package. The reference of the cointegration test is already mentioned (Pesaran et al, 2001). If you refer to the ARDL and ECM models, these are not usually cited as they are too common and developed gradually through time by many researchers.

---

**Statement of need:**

**Ha0ye:** _"Explain what kind of flexibility is encapsulated in ARDL and ECM. (the summary mentions they are very flexible, which is fine and good for modeling approaches, but what kind of situation warrants this flexibility?)"_

**Answer:** We don't understand what you mean by "what kind of flexibility". Flexibility is simply the capability of a model to fit complex relationships (e.g. not linear/straight lines). This can be achieved because of the autoregressive (AR) and the distributed lag (DL, essentially AR terms of the independent variables) terms in the ARDL model. So, should we consider this point as done?

**Ha0ye:** _"The software is designed to simplify the construction of these models (i.e. the user need only specify the order for the ARDL model). This should be emphasized as improving the accessibility of these modeling approaches for e.g. students and users who apply time series modeling, but are perhaps not experts in software, the mathematics of time series modeling, etc."_

**Answer:** We found this advise very useful and we updated the "Statement of need" accordingly.

**Ha0ye:** _"The link to the manual and recent example do not really fit here. Instead, link to a vignette with examples, and move the Qiu et al. to the summary."_

**Answer:** The link to the manual was used to point to where the detailed explanation of every function can be found, as this is one of the main points in the statement of need. Sure, this kind of information could also appear in a dedicated vignette instead of the manual, but vignettes are not necessary parts of R packages nor of JOSS submissions.
About the reference to Qiu et al. (2021), this is where the guidlines for authors and the example paper of JOSS suggests it should be.
Should we consider this point as ok?

---

**State of the field:**

**Ha0ye:** _"Mostly yes, but I think this would be improved by merging the content of the "Ease of use" section of the README here, as well as the "Conclusion" section of the software paper."_

**Answer:** Thank you for the suggestion. We merged the information of the "Ease of use" and added some more information regarding this section.

---

**Quality of writing:**

**Ha0ye:** _"The language is acceptable, though some expansion of the modeling approaches for the time series novice would be appreciated."_

**Answer:** We believe that describing the ARDL and the ECM equations mathematically is sufficient for understanding these models as they are very commonly used even in undergraduate level. More details were given before, but following your suggestions, for the violation of the software paper's guidelines, we deleted the detailed descriptions and rewrote that part.
Since we are a little bit confused about what you mean here, could you elaborate?

---

**References:**

**Ha0ye:** _"Citations should be added that expand on the time series methods."_

**Answer:** As we mentioned above about the "Summary", citations about the ARDL and ECM models would be very hard to be specified as they are very commonly used methods and their origins do not appear on specific papers.
However, we added Pesaran and Shin, (1999) which is the paper that connects these common models with the paper of focus (Pesaran et al., 2001).

---

**Additional Notes:**

**Answer:** Thank you for these suggestions! They are indeed very useful and we fixed some of them. `pkgdown` is in our next plans and for sure we will take a look at the CITATION.cff.
