### 2022-08-10 Natsiopoulos/ARDL#7 @Natsiopoulos
Hi @jessie-dotson
Thank you for reviewing our submission.
Here are our answers for your suggestions.

---

_"equations would be clearer if variables were defined"_

**Answer:** Most of the equations have been removed as suggested by the other reviewer.
Theses are general equation forms, and not a modeling application, so there are not much to define.
Unless you mean to refer to the constant and linear trend and the error, which we consider redundant to be mentioned for such a simple case.
Do you think we should mention these?

---

_"lines 50 - 54: Would be clearer if there were a clear explanation of what determined a "Case"_

**Answer:** In the first version of the software paper the cases where explicitly defined by the detailed equations. In this version, after the other reviewers suggestion about removing the detailed equations, we added the information that the case is about the restriction in the constant and the linear trend.
See in the software paper for "...the case for potential restriction in the deterministic parameters (constant and linear trend)..."

---

_"equations 4-8: ECT is not defined"_

**Answer:** These equations do not longer exist.

---

_"line 95: include link to API documentation when discussing the API"_

**Answer:** Thank you for the suggestion. We added the link as needed.
