### 2022-08-10 Natsiopoulos/ARDL#10 @Natsiopoulos
There are no automated test yet, although we have started working on that.
But there are already excessive tests for every function, (most of them included in the internal function `parsers` as this is the workhorse behind every other function). These are not in an automated format though, but rather manual tests.

As manual tests are permitted, can this be considered as done?

About the verification of the functionality, this has been verified through the recently published paper on the Journal of Applied Econometrics (we added this reference in the new version of the software paper).
