### 2022-08-10 Natsiopoulos/ARDL#8 @Natsiopoulos
It was just a warning that the object inside `as_tibble()` had no column names.
We fixed this by setting this object to `data.frame` instead of `matrix`, producing identical results.
