### 2022-08-10 Natsiopoulos/ARDL#6 @ha0ye
The edits look good! I followed up on a few points below:

> **Automated tests:**
>
> **Answer:** Thank you for suggesting using `testthat`, it is indeed something that we plan to do. However, there are already excessive tests for every function, (most of them included in the internal function `parsers` as this is the workhorse behind every other function). These are not in an automated format though, but rather manual tests.
>
> As manual tests are permitted, can this be considered as done?

Looking at the `parsers` function, you do seem to have extensive checking of inputs. However, the software testing that is expected is more like a correctness check. Usually of the form, given input X, the software produces result Y. This gives the user confidence that if they use the software, and the authors continue to evolve or improve on it, that they still have reproducible results with later versions. As mentioned in the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), the tests do not have to be automated, but you would need to specify some kind of examples of the inputs and expected outputs.

> **Summary:**
>
> **Ha0ye:** _"It is not clear with ARDL or ECM are, though it is implies that this package is implementing these approaches."_
>
> **Answer:** As these are widely known models in undergraduate level of various fields of studies (e.g. economics, statistics, environmental studies etc), we don't think that this should be overspecified. Also, ARDL is self defined by its name and we believe that a technical definition does not fit the purpose of this summary, as this is supposed to describe the "high-level functionality and purpose of the software for a diverse, non-specialist audience". Although, we added a reference (Pesaran and Shin, 1999) which is the paper that introduces the ARDL and ECM as platforms for the cointegration test in the paper of Pesaran et al, (2001). Please let us know if you think otherwise.

The review criteria expect that the summary be described "for a diverse, non-specialist audience." Maybe just add a phrase that these are time series models common in economics, statistics, environmental studies, etc. This way a reader who is unfamiliar with it can be sure that the software does not meet their needs, or has a bit of direction on how to search to find more background information.

> **Statement of need:**
>
> **Ha0ye:** _"Explain what kind of flexibility is encapsulated in ARDL and ECM. (the summary mentions they are very flexible, which is fine and good for modeling approaches, but what kind of situation warrants this flexibility?)"_
>
> **Answer:** We don't understand what you mean by "what kind of flexibility". Flexibility is simply the capability of a model to fit complex relationships (e.g. not linear/straight lines). This can be achieved because of the autoregressive (AR) and the distributed lag (DL, essentially AR terms of the independent variables) terms in the ARDL model. So, should we consider this point as done?

The description that you gave here can be included in the statement of need. My question hinges on the fact that models are described as "flexible" but with different meanings, such as robust to missing data, noise, probability distribution, etc. A clarification is helpful! (Again, emphasizing a bit more explanation for individuals not familiar with these models specifically.)
