### 2022-08-11 Natsiopoulos/ARDL#6 @Natsiopoulos
> The review criteria expect that the summary be described "for a diverse, non-specialist audience." Maybe just add a phrase that these are time series models common in economics, statistics, environmental studies, etc. This way a reader who is unfamiliar with it can be sure that the software does not meet their needs, or has a bit of direction on how to search to find more background information.

> The description that you gave here can be included in the statement of need. My question hinges on the fact that models are described as "flexible" but with different meanings, such as robust to missing data, noise, probability distribution, etc. A clarification is helpful! (Again, emphasizing a bit more explanation for individuals not familiar with these models specifically.)

We treated these issues adding the missing information.

> Looking at the `parsers` function, you do seem to have extensive checking of inputs. However, the software testing that is expected is more like a correctness check. Usually of the form, given input X, the software produces result Y. This gives the user confidence that if they use the software, and the authors continue to evolve or improve on it, that they still have reproducible results with later versions. As mentioned in the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), the tests do not have to be automated, but you would need to specify some kind of examples of the inputs and expected outputs.

Assuming that all the other issues are ok now, we will start working on building the tests and we will come back when they are ready.
