### 2022-08-17 f961ca9
- Added tests - Fixed bug in multipliers() (short-run without constant term) - Added BugReports & Video links in DESCRIPTION - Updated CITATION
