+++ tests/testthat/helper-functions.R
@@ -1,0 +1,7 @@
+AIC_pss <- function(model){
+    # maximized log-likelihood value of the model
+    LLp <- logLik(model)
+    # number of freely estimated coefficients
+    sp <- length(model$coefficients)
+    LLp - sp
+}
