### 2022-09-06 6e35af4
Reference update in JOSS paper
