+++ JOSS-paper/paper.bib
@@ -58,9 +58,10 @@
 author = {Natsiopoulos, Kleanthis and Tzeremes, Nickolaos G.},
 title = {ARDL bounds test for cointegration: Replicating the Pesaran et al. (2001) results for the UK earnings equation using R},
 journal = {Journal of Applied Econometrics},
-volume = {n/a},
-number = {n/a},
-pages = {},
+volume = {37},
+number = {5},
+year = {2022},
+pages = {1079-1090},
 keywords = {ARDL, bounds test, cointegration, earnings equation, R},
 doi = {https://doi.org/10.1002/jae.2919},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jae.2919},
