### 2022-09-06 Natsiopoulos/ARDL#6 @Natsiopoulos
Hi @ha0ye,
We created all the automated tests needed but there seems to be a problem that we can't fix yet.
This is found when we run `devtools::test()` and `devtools::check()`.
It is about the externally defined function `AIC_pss()` which we use in `test-auto_ardl.R`.
Do you have any idea how we can fix it?
