### 2022-09-08 Natsiopoulos/ARDL#6 @ha0ye
@Natsiopoulos I'm not sure. There's this - https://stackoverflow.com/questions/35941729/r-testthat-unit-test-data-and-helper-function-conventions

But I've always defined any additional variables or functions at the top of the individual test script that needs it.
