+++ R/helper-functions.R
@@ -1,6 +1,6 @@
 AIC_pss <- function(model){
     # maximized log-likelihood value of the model
-    LLp <- logLik(model)
+    LLp <- stats::logLik(model)
     # number of freely estimated coefficients
     sp <- length(model$coefficients)
     LLp - sp
