+++ README.Rmd
@@ -79,7 +79,7 @@ uecm_3132 <- uecm(ardl_3132)
 summary(uecm_3132)
 ```
 
-And also the RECM (Restricted Error Correction Model) of the underlying ARDL(3,1,3,2), allowing the constant to join the short-run relationship (case 2), instead of the long-run (case 3).
+And also the RECM (Restricted Error Correction Model) of the underlying ARDL(3,1,3,2), allowing the constant to join the long-run relationship (case 2), instead of the short-run (case 3).
 
 ```{r recm}
 recm_3132 <- recm(uecm_3132, case = 2)
