### 2022-09-08 Natsiopoulos/ARDL#6 @Natsiopoulos
@ha0ye - It seems that placing the function in the test files or placing the `helper-*.R` file in `testthat` folder is not an option anymore. It has to be as a non-exported function in the R folder, that's ok.
Now, what's the next step? Should I set a proper version number (curently it is 0.2.0.9000 indicating the development state)?
