### 2022-09-22 Natsiopoulos/ARDL#6 @ha0ye
Seems to work now. Maybe consider a release if JOSS approves the resubmission.

You could also look into add code coverage, to see where additional tests might be useful - https://github.com/r-lib/covr (100% coverage isn't needed, but you can verify that the important parts of the code are being tested).
