+++ README.md
@@ -4,6 +4,13 @@
 # ARDL <img src="man/figures/logo.png" align="right" width="120" />
 
 <!-- badges: start -->
+
+[![downloads](https://cranlogs.r-pkg.org/badges/grand-total/ARDL)](https://CRAN.R-project.org/package=ARDL)
+[![CRAN
+status](https://www.r-pkg.org/badges/version/ARDL)](https://CRAN.R-project.org/package=ARDL)
+[![Lifecycle:
+stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
+[![codecov](https://codecov.io/gh/Natsiopoulos/ARDL/branch/master/graph/badge.svg?token=PVZAY57F3D)](https://codecov.io/gh/Natsiopoulos/ARDL)
 <!-- badges: end -->
 
 ## Overview
