### 2022-09-24 f322cbc
Small corrections in JOSS paper
