### 2022-09-24 Natsiopoulos/ARDL#6 @Natsiopoulos
I checked the package coverage with covr and also connected codecov with the repo.
What's the next step now?
