### 2022-09-30 Natsiopoulos/ARDL#7 @jessie-dotson
Perhaps this is a difference in background -- as I warned when I agreed to review this submission, I don't use ARDL -- but I still believe the equations would be more useful if there were *some* pointers as to what the variables are that are used in equations 1 and 2.  Nonetheless, I'm not going ot pursue the point.
