### 2022-10-06 Natsiopoulos/ARDL#9 @Natsiopoulos
@jessie-dotson - The `BugReports` is in the `DESCRIPTION` file and the `CONTRIBUTING.md` is in the `.github` directory, where they should be.
