### 2022-10-06 Natsiopoulos/ARDL#9 @jessie-dotson
@Natsiopoulos  - Brilliant!  thank you for the finding chart.
