+++ JOSS-paper/paper.bib
@@ -47,7 +47,7 @@
   title = {{Maximum Likelihood Estimation and Inference on Cointegration – with Applications to the Demand for Money}},
   author = {Johansen, S. and Juselius, K.},
   journal = {Oxford Bulletin of Economics and Statistics},
-  doi = {https://doi.org/10.1111/j.1468-0084.1990.mp52002003.x},
+  doi = {10.1111/j.1468-0084.1990.mp52002003.x},
   number = {2},
   pages = {169--210},
   volume = {52},
@@ -63,7 +63,7 @@ number = {5},
 year = {2022},
 pages = {1079-1090},
 keywords = {ARDL, bounds test, cointegration, earnings equation, R},
-doi = {https://doi.org/10.1002/jae.2919},
+doi = {10.1002/jae.2919},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jae.2919},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jae.2919},
 abstract = {Summary This paper replicates the UK earnings equation using the autoregressive distributed lag (ARDL) modeling approach and the bounds test for cointegration by Pesaran et al. (Journal of Applied Econometrics, 2001, 16(3), 289–326). The findings from the narrow sense fully replicate the original results using the open-source language R and the ARDL package. In the wide sense replication, augmented data are employed, thus extending the end period from 1997:Q4 to 2019:Q4, using an alternative measure for union power. Adopting the new dataset, this study reinvestigates the UK earnings equation, thereby providing supporting evidence of a long-run relationship and reveals empirical findings about the long-run effects of productivity, unemployment, tax wedge, and union power on wages.}
@@ -108,6 +108,6 @@ collection={Econometric Society Monographs}
   pages={103155},
   year={2021},
   issn = {0160-7383},
-  doi = {https://doi.org/10.1016/j.annals.2021.103155},
+  doi = {10.1016/j.annals.2021.103155},
   publisher={Elsevier}
 }
