### 2022-10-07 #3496 @Natsiopoulos
@sbenthall I fixed the DOI's. The manuscript looks ok too.
