### 2022-10-07 #3496 @sbenthall
@Natsiopoulos Great. Next, can you please:

- Tag a release for publication
- Create and online archive with a DOI, for example using Zenodo
- Report the version number and archive DOI in this review thread
