### 2022-10-15 #3496 @Natsiopoulos
@sbenthall I made a minor change and tagged and released it as v0.2.1.
The DOI is 10.5281/zenodo.7200425
