### 2022-10-18 #3496 @sbenthall
@openjournals/joss-eics I recommend this paper for acceptance.

There seems to have been some metadata error, so I'm pinging you directly.
