### 2022-10-20 #3496 @arfon
@sbenthall – the error from @editorialbot is almost certainly around the fact that this BibTeX entry has the same ISSN and ISBN number (I think one of these is wrong): https://github.com/Natsiopoulos/ARDL/blob/master/JOSS-paper/paper.bib#L95-L96
