### 2022-10-20 #3496 @sbenthall
Thanks @arfon

@Natsiopoulos can you please:
 - Correct this problem with your references
 - Make a new archive and tag
 - Send me the new DOI and tag name
