+++ JOSS-paper/paper.bib
@@ -92,7 +92,6 @@ collection={Econometric Society Monographs}
   arxivId = {10.17863/CAM.5093},
   doi = {10.1002/jae.616},
   eprint = {CAM.5093},
-  isbn = {08837252},
   issn = {08837252},
   number = {3},
   pages = {289--326},
