### 2022-10-20 #3496 @Natsiopoulos
@sbenthall I corrected the references.
The new tag and release are v0.2.1 (I deleted the previous one and made a new one with the same name).
The new DOI is 10.5281/zenodo.7231155
