### 2022-10-31 #3496 @danielskatz
👋 @sbenthall & @Natsiopoulos - I'm glad to see this is close to done, with just a few actions left from @sbenthall and then @openjournals/sbcs-eics
