### 2022-10-31 #3496 @danielskatz
Just to be clear, @sbenthall should update the version and archive using editorialbot, then do the recommended accept again to make sure it fully works, and if the draft and everything else looks good, then notify @openjournals/sbcs-eics to complete the processing and acceptance
