### 2022-10-31 #3496 @Natsiopoulos
@sbenthall
Please consider this information for the tag, release, and DOI.

After a more detailed investigation, I made a minor correction in the paper, deleting half a sentence in order to be as fair as possible in comparison with the other packages.

The new tag and release are v0.2.1 (I deleted the previous one and made a new one with the same name).
The new DOI is 10.5281/zenodo.7267996
