### 2022-11-01 #3496 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
