### 2022-11-04 #3496 @Natsiopoulos
@arfon - At this point, is there something I need to do for the publication to be completed?
