### 2022-11-07 #3496 @arfon
> @arfon - At this point, is there something I need to do for the publication to be completed?

There is not. We're just wanting for the EiC for this area to pick up the paper and publish. Thanks for your patience!
