### 2022-11-08 #3496 @oliviaguest
@Natsiopoulos @sbenthall the DOI archive needs to have the same title and author list as the journal article. 😊
