### 2022-11-08 #3496 @oliviaguest
@sbenthall just FYI see the bullet points here to refresh your memory about DOI archives: https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
