### 2022-11-08 #3496 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
