### 2022-11-15 #3496 @Natsiopoulos
@oliviaguest @sbenthall Is there any problem with the acceptance procedure? I see that there are many recommendations for acceptance. Is there something that causes a problem at this step?
